requirements.txt
django~=3.0.0
pandas
python-telegram-bot